requirements.txt
streamlit==1.30.0
scikit-learn==1.3.2
imbalanced-learn==0.11.0
numpy==1.24.4
Pillow==10.0.0
